This tweaks the performance quite a bit.
We now wait for the threadpool to exit, rather then use a counter that indicates
if a task is running or not. This ended up being very contented with many cores and
at worst caused an extra 40% of scheduler overhead.

This also removes the ability to `extend` a task to multiple children. This had
questionable value, and it cost use two atomic adds which ended up being the 60%
of the execution time in add_task.